clean: split kubernetes-secrets client into seprate file + reorganise imports + reword error msgs
